recode pt4

+++ src/main/java/org/drachens/generalGame/faction/manage/DeleteCMD.java
@@ -2,6 +2,7 @@ package org.drachens.generalGame.faction.manage;
 
 import net.minestom.server.command.CommandSender;
 import net.minestom.server.command.builder.Command;
+import net.minestom.server.command.builder.arguments.Argument;
 import net.minestom.server.command.builder.arguments.ArgumentType;
 import net.minestom.server.command.builder.suggestion.SuggestionEntry;
 import org.drachens.Manager.defaults.ContinentalManagers;
@@ -15,11 +16,11 @@ public class DeleteCMD extends Command {
         setCondition((sender, s) -> leaderOfAFaction(sender));
         setDefaultExecutor((sender, context) -> sender.sendMessage("Proper usage: /faction delete <faction_name> "));
 
-        var factions = ArgumentType.String("factionName")
+        Argument<String> factions = ArgumentType.String("factionName")
                 .setSuggestionCallback((sender, context, suggestion) -> {
                     if (sender instanceof CPlayer player) {
                         Country country = player.getCountry();
-                        if (country != null) {
+                        if (null != country) {
                             if (country.isEconomyFactionLeader()) {
                                 suggestion.addEntry(new SuggestionEntry(country.getEconomy().getEconomyFactionType().getStringName()));
                             }
@@ -34,15 +35,14 @@ public class DeleteCMD extends Command {
             CPlayer player = (CPlayer) sender;
             Country country = player.getCountry();
             Faction factionToDelete = ContinentalManagers.world(player.getInstance()).countryDataManager().getFaction(context.get(factions));
-            if (factionToDelete != null && factionToDelete.isLeader(country))
-                factionToDelete.delete();
+            if (null != factionToDelete && factionToDelete.isLeader(country)) factionToDelete.delete();
         }, factions);
     }
 
     private boolean leaderOfAFaction(CommandSender sender) {
         if (sender instanceof CPlayer player) {
             Country country = player.getCountry();
-            return country != null && country.isLeaderOfAFaction();
+            return null != country && country.isLeaderOfAFaction();
         }
         return false;
     }
